requirements.txt
pandas==1.0.5
statsmodels==0.11.1
biopython==1.78